refactor(backend): arrange main.ts
