+++ backend/src/main.ts
@@ -1,25 +1,12 @@
-import { ValidationPipe } from "@nestjs/common";
-import { NestFactory } from "@nestjs/core";
-import {
-  FastifyAdapter,
-  NestFastifyApplication,
-} from "@nestjs/platform-fastify";
-import { DocumentBuilder, SwaggerModule } from "@nestjs/swagger";
-import { AppModule } from "./app.module";
+import { SwaggerModule } from "@nestjs/swagger";
+import { createApp } from "./app/app";
+import { generateOpenAPI } from "./app/openapi";
 
 async function bootstrap() {
-  const app = await NestFactory.create<NestFastifyApplication>(
-    AppModule,
-    new FastifyAdapter(),
-  );
-  app.useGlobalPipes(new ValidationPipe({ transform: true }));
+  const app = await createApp();
 
-  // Swagger
-  const config = new DocumentBuilder()
-    .setTitle(process.env.npm_package_name ?? "backend")
-    .setVersion(process.env.npm_package_version ?? "unknown")
-    .build();
-  const document = SwaggerModule.createDocument(app, config);
+  const document = generateOpenAPI(app);
+
   SwaggerModule.setup("spec", app, document);
 
   await app.listen(3100, "0.0.0.0");
