+++ backend/test/app.e2e-spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from "@nestjs/testing";
 import { INestApplication } from "@nestjs/common";
 import * as request from "supertest";
-import { AppModule } from "../src/app.module";
+import { AppModule } from "../src/app/app.module";
 
 describe("AppController (e2e)", () => {
   let app: INestApplication;
